Updated versions of nodes.dmp and names.dmp.
Removed 'cellular organisms' from both files. Maintained 'root' and 'Archae' and 'Bacteria' are children nodes of 'root'.
Third field in names.dmp is now empty since it only had a repetition of the second field.
Species names have been corrected so that the full name (binomial nomenclature) is used, instead of only the epithet.
Due to the changes the identifiers have changed.